Refactor :: WorkspaceEntity 를 Entity 폴더 아래로 수정함 #68
