+++ src/main/kotlin/com/seugi/api/domain/workspace/domain/WorkspaceRepository.kt
@@ -1,5 +1,6 @@
 package com.seugi.api.domain.workspace.domain
 
+import com.seugi.api.domain.workspace.domain.entity.WorkspaceEntity
 import org.springframework.data.repository.CrudRepository
 
 interface WorkspaceRepository: CrudRepository<WorkspaceEntity, Long> {
